python/requirements.txt
tensorflow-gpu==2.2.0
Flask==1.1.2
efficientnet==1.1.0
injector==0.18.4
Flask-Executor==0.9.4
Flask-Injector==0.12.3
Flask-Script==2.0.6
Flask-Bootstrap==3.3.7.1
dependency-injector
pyyaml
pytest-flask
pytest-cov